+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import './secondview.dart';
 
 void main() {
   runApp(const MyApp());
@@ -116,55 +117,4 @@ Widget _item(text) {
     padding: const EdgeInsets.all(20.0),
     child: Text(text, style: const TextStyle(fontSize: 18)),
   );
-}
-
-Widget _insertTaskField() {
-  return Container(
-      margin: const EdgeInsets.only(left: 16, right: 16),
-      child: const TextField(
-        decoration: InputDecoration(
-          hintText: 'What are you going to do?',
-          focusedBorder: OutlineInputBorder(
-            borderSide: BorderSide(color: Colors.black, width: 3.0),
-          ),
-        ),
-      ));
-}
-
-Widget _addtext() {
-  return Row(mainAxisAlignment: MainAxisAlignment.center, children: [
-    ElevatedButton.icon(
-      onPressed: () {},
-      style: ElevatedButton.styleFrom(
-        backgroundColor: Colors.grey,
-        foregroundColor: Colors.black,
-      ),
-      icon: const Icon(
-        Icons.add,
-        size: 30.0,
-      ),
-      label: const Text('ADD'),
-    ),
-  ]);
-}
-
-class SecondView extends StatelessWidget {
-  const SecondView({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          title: const Text('TIG169 TODO'),
-        ),
-        body: Center(
-            child: Column(
-          mainAxisAlignment: MainAxisAlignment.start,
-          children: [
-            Container(height: 50),
-            SizedBox(height: 100, child: _insertTaskField()),
-            _addtext(),
-          ],
-        )));
-  }
 }
